src/main/java/ru/netology/Radio.java
package ru.netology;

public class Radio {
    private int currentStation;


    public int getCurrentStation() {
        return currentStation;
    }

    public void setCurrentStation(int newCurrentStation) {
        if (newCurrentStation < 0) {
            return;
        }
        if (newCurrentStation > 9) {
            return;
        }
        currentStation = newCurrentStation;

    }

    public void nextStation() {
        if (9 > currentStation) {
            setCurrentStation(currentStation + 1);
        } else {

            setCurrentStation(0);
        }
    }

    public void prevStation() {
        if (0 < currentStation) {
            setCurrentStation(currentStation - 1);
        } else {
            setCurrentStation(9);
        }
    }

    private int currentVolume;

    public int getCurrentVolume() {
        return currentVolume;
    }

    public void setCurrentVolume(int newCurrentVolume) {
        if (newCurrentVolume < 0) {
            return;
        }
        if (newCurrentVolume > 10) {
            return;
        }
        currentVolume = newCurrentVolume;

    }

    public void increaseVolume() {
        if (currentVolume < 10) {
            currentVolume = currentVolume + 1;
        }

    }

    public void decreaseVolume() {
        if (currentVolume > 0) {
            currentVolume = currentVolume - 1;
        }

    }


}
